+++ ktor-network/posix/src/io/ktor/network/sockets/CIOWriter.kt
@@ -30,7 +30,7 @@ internal fun CoroutineScope.attachForWritingImpl(
                 val bufferStart = pinned.addressOf(start).reinterpret<ByteVar>()
                 val remaining = stop - start
                 val bytesWritten = if (remaining > 0) {
-                    ktor_send(descriptor, bufferStart, remaining.convert(), 0).toInt()
+                    ktor_send(descriptor, bufferStart, remaining.convert(), 0)
                 } else {
                     0
                 }
